Organize code and refactor the way custom stats are sent to the client
One thing to note is that the way of sending custom stats works
differently in 1.16.5, 1.17.1, and 1.18.2 when the mod is on the client and
server
In 1.16.5, nothing is sent to the client.
In 1.17.1, the names of the custom stats are sent, but the values are
not.
In 1.18.2, both the names and values of custom stats are sent.